Added in a move counter & and end goal
Added in a move counter UI element that counts down number of available moves and an end goal Gameobject that sends you to the next level when you touch it.

Running out of moves brings up a popup to restart the level. Implementation of this was hacky as fuck but I couldn't think of a better way to do it.